[CHORE] Move CuArrayVec from cu_apriltag to cu29_runtime
