+++ core/cu29_runtime/src/payload.rs
@@ -2,6 +2,10 @@
 ///
 /// The constraint on the messages is that they can be part of a copper list, fixed sized and bincode serializable.
 use arrayvec::ArrayVec;
+use bincode::de::{BorrowDecoder, Decoder};
+use bincode::enc::Encoder;
+use bincode::error::{DecodeError, EncodeError};
+use bincode::BorrowDecode;
 use bincode::{Decode, Encode};
 
 /// Copper friendly wrapper for a fixed size array.
@@ -86,5 +90,71 @@ where
         }
 
         Ok(Self { inner })
+    }
+}
+
+/// TODO: Move that to the runtime as it is useful.
+#[derive(Debug, Clone)]
+pub struct CuArrayVec<T, const N: usize>(pub ArrayVec<T, N>);
+
+impl<T, const N: usize> Default for CuArrayVec<T, N> {
+    fn default() -> Self {
+        Self(ArrayVec::new())
+    }
+}
+
+impl<T, const N: usize> Encode for CuArrayVec<T, N>
+where
+    T: Encode + 'static,
+{
+    fn encode<E: Encoder>(&self, encoder: &mut E) -> Result<(), EncodeError> {
+        let CuArrayVec(inner) = self;
+        inner.as_slice().encode(encoder)
+    }
+}
+
+impl<T, const N: usize> Decode<()> for CuArrayVec<T, N>
+where
+    T: Decode<()> + 'static,
+{
+    fn decode<D: Decoder<Context = ()>>(decoder: &mut D) -> Result<Self, DecodeError> {
+        let inner = Vec::<T>::decode(decoder)?;
+        let actual_len = inner.len();
+        if actual_len > N {
+            return Err(DecodeError::ArrayLengthMismatch {
+                required: N,
+                found: actual_len,
+            });
+        }
+
+        let mut array_vec = ArrayVec::new();
+        for item in inner {
+            array_vec.push(item); // Push elements one by one
+        }
+        Ok(CuArrayVec(array_vec))
+    }
+}
+
+impl<'de, T, const N: usize> BorrowDecode<'de, ()> for CuArrayVec<T, N>
+where
+    T: BorrowDecode<'de, ()> + 'static,
+{
+    fn borrow_decode<D: BorrowDecoder<'de, Context = ()>>(
+        decoder: &mut D,
+    ) -> Result<Self, DecodeError> {
+        let inner = Vec::<T>::borrow_decode(decoder)?;
+        let actual_len = inner.len();
+        if actual_len > N {
+            return Err(DecodeError::ArrayLengthMismatch {
+                required: N,
+                found: actual_len,
+            });
+        }
+
+        let mut array_vec = ArrayVec::new();
+        for item in inner {
+            array_vec.push(item); // Push elements one by one
+        }
+        Ok(CuArrayVec(array_vec))
     }
 }
